all function imports now correct in  temul.api; test: old import exceptions
