+++ temul/polarisation.py
@@ -1,10 +1,7 @@
 
-raise DeprecationWarning("\n "
-      "WARNING: "
-      "The temul.polarisation module is deprecated and will be removed in "
-      "future versions of the TEMUL Toolkit. To use the polarisation "
-      "functionality please use `import temul.topotem as tt`.\n"
-      "WARNING: "
-      "The polarisation module will be removed from the TEMUL Toolkit in "
-      "future versions. Its functionality will be moved to the TopoTEM Python "
-      "Package.")
+raise DeprecationWarning(
+      "\n WARNING: "
+      "To use the polarisation "
+      "functionality please use `import temul.api as tml`. "
+      "The temul.polarisation module is deprecated and is now "
+      "`temul.topotem.polarisation.\n")
