+++ temul/topotem/__init__.py
@@ -1,6 +1,7 @@
 
 from .fft_mapping import (
     choose_mask_coordinates, get_masked_ifft,
+    choose_points_on_image,
 )
 
 from .polarisation import (
